% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_labels}
\alias{example_labels}
\title{Example Cell Labels}
\format{
A character vector of length 49.
}
\source{
Biase et al. single-cell RNA-seq dataset, NCBI Gene Expression Omnibus,
accession GSE57249.
}
\usage{
example_labels
}
\description{
Cell identity labels corresponding to the 49 columns of
\code{example_data}. The labels are derived from the cell identities
established in the Biase et al. study associated with GEO accession
GSE57249.
}
\details{
For use in sclValid, the original cell groups were relabeled as \code{"A"},
\code{"B"}, and \code{"C"}.
}
\keyword{datasets}
