% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_validation.R
\name{make_sce}
\alias{make_sce}
\title{Create a SingleCellExperiment Object}
\usage{
make_sce(x, labels)
}
\arguments{
\item{x}{A numeric matrix with features in rows and observations in columns.}

\item{labels}{A vector or factor containing one label for each column of \code{x}.}
}
\value{
A \code{SingleCellExperiment} object containing the input matrix and
observation labels.
}
\description{
Creates a \code{SingleCellExperiment} object from a numeric feature matrix and
a corresponding vector of observation labels. The input matrix is stored
as the \code{"data"} assay and the supplied labels are stored in \code{colData}.
}
\examples{
sce <- make_sce(example_data, example_labels)
sce

}
