% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_cidr.R
\name{run_cidr}
\alias{run_cidr}
\title{Run CIDR Clustering}
\usage{
run_cidr(sce, cluster_sizes, verbose = TRUE)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object containing a \code{"data"} assay.}

\item{cluster_sizes}{An integer vector specifying the numbers of clusters
to fit.}

\item{verbose}{Logical indicating whether detailed progress messages should
be displayed during clustering`.}
}
\value{
The input \code{SingleCellExperiment} object with additional CIDR
cluster-assignment columns in \code{colData}.
}
\description{
Applies the CIDR clustering procedure to the data assay of a
\code{SingleCellExperiment} object for one or more requested numbers of clusters.
Cluster assignments are added to the object's \code{colData}.
}
\details{
For a cluster size \code{k}, the resulting column is named \code{"CIDR_cs{k}"}.
}
\examples{
sce <- make_sce(example_data, example_labels)
sce <- run_cidr(sce, cluster_sizes = 2:4)
SummarizedExperiment::colData(sce)

}
\references{
Lin P, Troup M, Ho JWK. CIDR: Ultrafast and accurate clustering through
imputation for single-cell RNA-seq data.
Genome Biology. 2017;18:59.
}
