% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_pcareduce.R
\name{run_pcareduce}
\alias{run_pcareduce}
\title{Run pcaReduce Clustering}
\usage{
run_pcareduce(sce, cluster_sizes, method = "M", nbt = 100, verbose = TRUE)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object containing a \code{"data"} assay.}

\item{cluster_sizes}{An integer vector specifying the numbers of clusters
to fit.}

\item{method}{Character string specifying the pcaReduce merging method.
Must be either \code{"M"} or \code{"S"}.}

\item{nbt}{Number of pcaReduce iterations to perform. Defaults to \code{100}.}

\item{verbose}{Logical indicating whether detailed progress messages should
be displayed during clustering`.}
}
\value{
The input \code{SingleCellExperiment} object with additional pcaReduce
cluster-assignment columns in \code{colData}.
}
\description{
Applies the pcaReduce clustering procedure to the data assay of a
\code{SingleCellExperiment} object for one or more requested numbers of clusters.
Cluster assignments are added to the object's \code{colData}.
}
\details{
The pcaReduce method may be run using either the \code{"M"} or \code{"S"} merging
procedure. For a cluster size \code{k}, resulting columns are named
\code{"pcaReduceM_cs{k}"} or \code{"pcaReduceS_cs{k}"}, depending on the selected
method.
}
\examples{
sce <- make_sce(example_data, example_labels)
sce <- run_pcareduce(
  sce,
  cluster_sizes = 2:4,
  method = "M"
)
SummarizedExperiment::colData(sce)

}
\references{
Žurauskienė J, Yau C.
pcaReduce: Hierarchical clustering of single cell transcriptional profiles.
BMC Bioinformatics. 2016;17:140.
}
