% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclValid-package.R
\docType{package}
\name{sclValid-package}
\alias{sclValid}
\alias{sclValid-package}
\title{sclValid}
\description{
Provides methods for evaluating and comparing clustering results, with an emphasis on sample-level clustering where observations are treated as the units being clustered and perturbation is performed by removing samples. The package implements internal, stability, and external validation measures and supports the comparison and ranking of clustering procedures across tuning parameter choices. It is conceptually related to the 'clValid' package, which assesses clustering stability by perturbing features; in contrast, this package focuses on perturbing the samples being clustered, making the validation framework suitable for applications such as single-cell clustering where samples or cells are the primary units of interest. Methods are based in part on Visser and Datta (2025) \doi{10.1002/sim.70331}.
}
\author{
\strong{Maintainer}: Owen Visser \email{owviss@gmail.com}

Authors:
\itemize{
  \item Owen Visser \email{owviss@gmail.com}
  \item Somnath Datta [contributor]
}

Other contributors:
\itemize{
  \item Justina Žurauskienė (pcaReduce compatibility code adapted from the pcaReduce package) [contributor]
  \item Christopher Yau (pcaReduce compatibility code adapted from the pcaReduce package) [contributor]
}

}
\keyword{internal}
