% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_apn.R
\name{validation_apn}
\alias{validation_apn}
\title{Calculate Average Proportion of Non-Overlap}
\usage{
validation_apn(clusters, reduced_clusters, index)
}
\arguments{
\item{clusters}{A vector containing cluster assignments obtained from the
full dataset.}

\item{reduced_clusters}{A list of clustering-assignment vectors obtained
from the reduced datasets.}

\item{index}{A list specifying the observations removed for each reduced
clustering solution.}
}
\value{
A numeric average proportion of non-overlap. Lower values indicate
more stable clustering.
}
\description{
Calculates an adapted Average Proportion of Non-Overlap (APN) stability
measure by comparing cluster memberships between full and perturbed
clustering solutions.
}
\details{
The implementation is adapted from the clustering stability measures
described by Datta and Datta.
}
\examples{
clusters <- c(1, 1, 2, 2)

reduced_clusters <- list(
  c(1, 2, 2),
  c(1, 2, 2),
  c(1, 1, 2),
  c(1, 1, 2)
)

index <- as.list(1:4)

validation_apn(
  clusters,
  reduced_clusters,
  index
)

}
\references{
Adapted from:
Datta S, Datta S. Comparisons and validation of statistical clustering
techniques for microarray gene expression data.
Bioinformatics. 2003;19(4):459-466.
}
