% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_ari.R
\name{validation_ari}
\alias{validation_ari}
\title{Calculate Adjusted Rand Index}
\usage{
validation_ari(clusters, labels)
}
\arguments{
\item{clusters}{A vector containing one cluster assignment per observation.}

\item{labels}{A vector containing the known label for each observation.}
}
\value{
A numeric adjusted Rand index.
}
\description{
Calculates the adjusted Rand index between a clustering solution and known
observation labels. The adjusted Rand index measures agreement between two
partitions while correcting for agreement expected by chance.
}
\examples{
clusters <- c(1, 1, 2, 2)
labels <- c("A", "A", "B", "B")
validation_ari(clusters, labels)

}
\references{
Hubert L, Arabie P. Comparing partitions.
Journal of Classification. 1985;2:193-218.
}
