% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_average_distance_means.R
\name{validation_average_distance_means}
\alias{validation_average_distance_means}
\title{Calculate Average Distance Between Means}
\usage{
validation_average_distance_means(clusters, reduced_clusters, data, index)
}
\arguments{
\item{clusters}{A vector containing cluster assignments obtained from the
full dataset.}

\item{reduced_clusters}{A list of clustering-assignment vectors obtained
from the reduced datasets.}

\item{data}{A numeric matrix with features in rows and observations in
columns.}

\item{index}{A list specifying the observations removed for each reduced
clustering solution.}
}
\value{
A numeric average-distance-between-means stability score. Lower
values indicate greater stability in cluster locations.
}
\description{
Calculates an adapted Average Distance Between Means (ADM) stability
measure based on distances between cluster centroids in full and perturbed
clustering solutions.
}
\details{
The implementation is adapted from the clustering stability measures
described by Datta and Datta.
}
\examples{
clusters <- c(1, 1, 2, 2)
x <- matrix(
  c(
    0, 0, 3, 3,
    0, 1, 3, 4
  ),
  nrow = 2,
  byrow = TRUE
)

reduced_clusters <- list(
  c(1, 2, 2),
  c(1, 2, 2),
  c(1, 1, 2),
  c(1, 1, 2)
)

index <- as.list(1:4)

validation_average_distance_means(
  clusters,
  reduced_clusters,
  x,
  index
)

}
\references{
Adapted from:
Datta S, Datta S. Comparisons and validation of statistical clustering
techniques for microarray gene expression data.
Bioinformatics. 2003;19(4):459-466.
}
