% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_bhi.R
\name{validation_bhi}
\alias{validation_bhi}
\title{Calculate Biological Homogeneity Index}
\usage{
validation_bhi(clusters, labels)
}
\arguments{
\item{clusters}{A vector containing one cluster assignment per observation.}

\item{labels}{A vector containing the known label for each observation.}
}
\value{
A numeric biological homogeneity index between zero and one, with
larger values indicating greater biological homogeneity.
}
\description{
Calculates the Biological Homogeneity Index (BHI), which measures
homogeneity of known biological labels within clusters.
}
\examples{
clusters <- c(1, 1, 2, 2)
labels <- c("A", "A", "B", "B")
validation_bhi(clusters, labels)

}
\references{
Pihur V, Brock GN, Datta S.
Cluster validation for microarray data: an appraisal.
In: Advances in Multivariate Statistical Methods.
World Scientific; 2009:79-94.
}
