% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclValid-package.R
\docType{package}
\name{sclValid-package}
\alias{sclValid}
\alias{sclValid-package}
\title{sclValid}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
