% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclValid-package.R
\docType{package}
\name{sclValid-package}
\alias{sclValid}
\alias{sclValid-package}
\title{sclValid}
\description{
Provides tools for clustering, validation, and ranking of single-cell RNA sequencing data. The package implements multiple clustering approaches, a collection of internal, external, and stability-based validation measures, and rank aggregation for comparing clustering solutions. User-defined clustering functions can also be incorporated into the validation framework. Methods are based in part on Visser and Datta (2025) [doi:10.1002/sim.70331](doi:10.1002/sim.70331).
}
\author{
\strong{Maintainer}: Owen Visser \email{owviss@gmail.com}

Authors:
\itemize{
  \item Owen Visser \email{owviss@gmail.com}
}

Other contributors:
\itemize{
  \item Justina Žurauskienė (pcaReduce compatibility code adapted from the pcaReduce package) [contributor]
  \item Christopher Yau (pcaReduce compatibility code adapted from the pcaReduce package) [contributor]
}

}
\keyword{internal}
