% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclValid-package.R
\docType{package}
\name{sclValid-package}
\alias{sclValid}
\alias{sclValid-package}
\title{sclValid}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Owen Visser \email{owviss@gmail.com}

Authors:
\itemize{
  \item Owen Visser \email{owviss@gmail.com}
}

Other contributors:
\itemize{
  \item Justina Žurauskienė (pcaReduce compatibility code adapted from the pcaReduce package) [contributor]
  \item Christopher Yau (pcaReduce compatibility code adapted from the pcaReduce package) [contributor]
}

}
\keyword{internal}
